+++ src/EntityAliasDependencyCollectionInterface.php
@@ -7,13 +7,9 @@ use Drupal\Core\Entity\EntityInterface;
 
 interface EntityAliasDependencyCollectionInterface
 {
-    public function getAliases(): array;
-
     public function getEntities(): array;
 
     public function getConfigs(): array;
-
-    public function addPathAlias(string $pid): void;
 
     public function addEntity(EntityInterface $entity): void;