+++ src/EntityAliasDependencyInterface.php
@@ -8,7 +8,6 @@ interface EntityAliasDependencyInterface
 {
     public const TYPE_ENTITY = 'entity';
     public const TYPE_CONFIG = 'config';
-    public const TYPE_PATH_ALIAS = 'pid';
 
     public function getDependentEntity(): ?EntityInterface;