+++ src/PatternTokenDependencyProviderBase.php
@@ -5,8 +5,11 @@ namespace Drupal\wmpathauto;
 use Drupal\Component\Plugin\PluginBase;
 use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Entity\EntityMalformedException;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
+use Drupal\Core\Entity\FieldableEntityInterface;
 use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
 use Drupal\Core\Utility\Token;
+use Drupal\path_alias\PathAliasInterface;
 use Drupal\pathauto\AliasStorageHelperInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -14,6 +17,8 @@ abstract class PatternTokenDependencyProviderBase extends PluginBase implements 
 {
     /** @var Token */
     protected $tokens;
+    /** @var EntityTypeManagerInterface */
+    protected $entityTypeManager;
     /** @var AliasStorageHelperInterface */
     protected $aliases;
     /** @var PatternTokenDependencyProviderManager */
@@ -26,6 +31,7 @@ abstract class PatternTokenDependencyProviderBase extends PluginBase implements 
     ) {
         $instance = new static($configuration, $plugin_id, $plugin_definition);
         $instance->tokens = $container->get('token');
+        $instance->entityTypeManager = $container->get('entity_type.manager');
         $instance->aliases = $container->get('pathauto.alias_storage_helper');
         $instance->manager = $container->get('plugin.manager.pattern_token_dependency_provider');
 
@@ -39,21 +45,27 @@ abstract class PatternTokenDependencyProviderBase extends PluginBase implements 
             ->addDependencies($tokens, $data, $options, $dependencies);
     }
 
-    /** TODO: Get from path field */
-    protected function getEntityAlias(EntityInterface $entity): ?array
+    protected function getPathAliasByEntity(EntityInterface $entity): ?PathAliasInterface
     {
         try {
-            $source = '/' . $entity->toUrl()->getInternalPath();
+            return $this->getPathAlias(
+                '/' . $entity->toUrl()->getInternalPath(),
+                $entity->language()->getId()
+            );
         } catch (EntityMalformedException $e) {
             return null;
         }
+    }
 
-        $language = $entity->language()->getId();
+    protected function getPathAlias(string $path, string $langcode): ?PathAliasInterface
+    {
+        $entities = $this->entityTypeManager
+            ->getStorage('path_alias')
+            ->loadByProperties([
+                'langcode' => $langcode,
+                'path' => $path,
+            ]);
 
-        if ($alias = $this->aliases->loadBySource($source, $language)) {
-            return $alias;
-        }
-
-        return null;
+        return array_pop($entities);
     }
 }