+++ src/Plugin/PatternTokenDependencyProvider/MenuLink.php
@@ -69,8 +69,8 @@ class MenuLink extends PatternTokenDependencyProviderBase
                 $dependencies->addEntity($linkEntity);
 
                 $referencedEntity = $this->getReferencedEntity($linkPlugin, $langcode);
-                if ($referencedEntity && $alias = $this->getEntityAlias($referencedEntity)) {
-                    $dependencies->addPathAlias($alias['pid']);
+                if ($referencedEntity && $alias = $this->getPathAliasByEntity($referencedEntity)) {
+                    $dependencies->addEntity($alias);
                 }
             }